@@ -25,36 +25,41 @@ class SingleDropdownRange extends Component {
 	constructor(props) {
 		super(props);
 
+		const defaultValue = props.defaultValue || props.value;
+		const value = props.selectedValue || defaultValue || null;
+		const currentValue = SingleDropdownRange.parseValue(value, props);
+
 		this.state = {
-			currentValue: null,
+			currentValue,
 		};
 		this.type = 'range';
 		this.locked = false;
+
+		props.addComponent(props.componentId);
 		props.setQueryListener(props.componentId, props.onQueryChange, null);
-	}
 
-	componentWillMount() {
-		this.props.addComponent(this.props.componentId);
-		this.setReact(this.props);
+		this.setReact(props);
+		const hasMounted = false;
 
-		if (this.props.selectedValue) {
-			this.setValue(this.props.selectedValue, true);
-		} else if (this.props.defaultSelected) {
-			this.setValue(this.props.defaultSelected, true);
+		if (currentValue) {
+			this.setValue(currentValue.label, true, props, hasMounted);
 		}
 	}
 
-	componentWillReceiveProps(nextProps) {
-		checkPropChange(this.props.react, nextProps.react, () => this.setReact(nextProps));
+	componentDidUpdate(prevProps) {
+		checkPropChange(this.props.react, prevProps.react, () => this.setReact(this.props));
 
-		checkPropChange(this.props.dataField, nextProps.dataField, () => {
-			this.updateQuery(this.state.currentValue, nextProps);
+		checkPropChange(this.props.dataField, prevProps.dataField, () => {
+			this.updateQuery(this.state.currentValue, prevProps);
 		});
 
-		if (!isEqual(this.props.defaultSelected, nextProps.defaultSelected)) {
-			this.setValue(nextProps.defaultSelected, true);
-		} else if (!isEqual(this.state.currentValue, nextProps.selectedValue)) {
-			this.setValue(nextProps.selectedValue, true);
+		if (!isEqual(this.props.value, prevProps.value)) {
+			this.setValue(this.props.value, true);
+		} else if (
+			!isEqual(this.state.currentValue, this.props.selectedValue)
+			&& !isEqual(this.props.selectedValue, prevProps.selectedValue)
+		) {
+			this.setValue(this.props.selectedValue || null, true);
 		}
 	}
 
@@ -86,7 +91,7 @@ class SingleDropdownRange extends Component {
 		return null;
 	};
 
-	setValue = (value, isDefaultValue = false, props = this.props) => {
+	setValue = (value, isDefaultValue = false, props = this.props, hasMounted = true) => {
 		// ignore state updates when component is locked
 		if (props.beforeValueChange && this.locked) {
 			return;
@@ -95,21 +100,26 @@ class SingleDropdownRange extends Component {
 		this.locked = true;
 		let currentValue = value;
 		if (isDefaultValue) {
-			currentValue = props.data.find(item => item.label === value) || null;
 			currentValue = SingleDropdownRange.parseValue(value, props);
 		}
 
 		const performUpdate = () => {
-			this.setState(
-				{
-					currentValue,
-				},
-				() => {
-					this.updateQuery(currentValue, props);
-					this.locked = false;
-					if (props.onValueChange) props.onValueChange(currentValue);
-				},
-			);
+			const handleUpdates = () => {
+				this.updateQuery(currentValue, props);
+				this.locked = false;
+				if (props.onValueChange) props.onValueChange(currentValue);
+			};
+
+			if (hasMounted) {
+				this.setState(
+					{
+						currentValue,
+					},
+					handleUpdates,
+				);
+			} else {
+				handleUpdates();
+			}
 		};
 
 		checkValueChange(props.componentId, currentValue, props.beforeValueChange, performUpdate);
@@ -129,6 +139,15 @@ class SingleDropdownRange extends Component {
 		});
 	};
 
+	handleChange = (selectedItem) => {
+		const { value, onChange } = this.props;
+		if (value) {
+			if (onChange) onChange(selectedItem);
+		} else {
+			this.setValue(selectedItem);
+		}
+	};
+
 	render() {
 		return (
 			<Container style={this.props.style} className={this.props.className}>
@@ -140,7 +159,7 @@ class SingleDropdownRange extends Component {
 				<Dropdown
 					innerClass={this.props.innerClass}
 					items={this.props.data}
-					onChange={this.setValue}
+					onChange={this.handleChange}
 					selectedItem={this.state.currentValue}
 					placeholder={this.props.placeholder}
 					keyField="label"
@@ -166,11 +185,13 @@ SingleDropdownRange.propTypes = {
 	customQuery: types.func,
 	data: types.data,
 	dataField: types.stringRequired,
-	defaultSelected: types.string,
+	defaultValue: types.string,
+	value: types.string,
 	filterLabel: types.string,
 	innerClass: types.style,
 	onQueryChange: types.func,
 	onValueChange: types.func,
+	onChange: types.func,
 	placeholder: types.string,
 	react: types.react,
 	showFilter: types.bool,
